Move App.js contents to user.js
