+++ mudae-showcase/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import Auth from './components/Auth'
+import Auth from './components/auth'
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
